+++ app/(auth)/layout.tsx
@@ -1,4 +1,5 @@
 import { Suspense } from "react";
+import Image from "next/image";
 
 export default function RootLayout({
     children,
@@ -8,6 +9,13 @@ export default function RootLayout({
     return (
         <>
             <div className='w-full flex flex-row justify-center align-middle p-16'>
+                <Image
+                    className='fixed bottom-8'
+                    src='/internbot-transparent.png'
+                    width={96}
+                    height={96}
+                    alt='Internbot Logo'
+                />
                 <Suspense fallback={<div>Loading...</div>}>{children}</Suspense>
             </div>
         </>
