+++ app/globals.css
@@ -4,8 +4,8 @@
 
 :root {
     --foreground-rgb: 0, 0, 0;
-    --background-start-rgb: 214, 219, 220;
-    --background-end-rgb: 255, 255, 255;
+    --background-start-rgb: 74, 74, 74;
+    --background-end-rgb: 55, 55, 55;
 }
 
 body {
